Make detours common for dup, dup2, dup3, & fcntl
