+++ mirrord-layer/src/socket/hooks.rs
@@ -235,105 +235,6 @@ pub(super) unsafe extern "C" fn uv__accept4_detour(
     accept4_detour(sockfd, address, address_len, flags)
 }
 
-/// https://github.com/metalbear-co/mirrord/issues/184
-#[hook_fn]
-pub(super) unsafe extern "C" fn fcntl_detour(fd: c_int, cmd: c_int, mut arg: ...) -> c_int {
-    trace!("fcntl_detour -> fd {:#?} | cmd {:#?}", fd, cmd);
-
-    let arg = arg.arg::<usize>();
-    let fcntl_result = FN_FCNTL(fd, cmd, arg);
-    let guard = crate::DetourGuard::new();
-    if guard.is_none() {
-        return fcntl_result;
-    }
-
-    if fcntl_result == -1 {
-        fcntl_result
-    } else {
-        let (Ok(result) | Err(result)) = fcntl(fd, cmd, fcntl_result)
-            .map(|()| fcntl_result)
-            .map_err(From::from);
-
-        trace!("fcntl_detour -> result {:#?}", result);
-        result
-    }
-}
-
-#[hook_guard_fn]
-pub(super) unsafe extern "C" fn dup_detour(fd: c_int) -> c_int {
-    trace!("dup_detour -> fd {:#?}", fd);
-
-    let dup_result = FN_DUP(fd);
-
-    if dup_result == -1 {
-        dup_result
-    } else {
-        let (Ok(result) | Err(result)) =
-            dup(fd, dup_result)
-                .map(|()| dup_result)
-                .map_err(|fail| match fail {
-                    LayerError::LocalFDNotFound(_) => dup_result,
-                    _ => fail.into(),
-                });
-
-        trace!("dup_detour -> result {:#?}", result);
-        result
-    }
-}
-
-#[hook_guard_fn]
-pub(super) unsafe extern "C" fn dup2_detour(oldfd: c_int, newfd: c_int) -> c_int {
-    trace!("dup2_detour -> oldfd {:#?} | newfd {:#?}", oldfd, newfd);
-
-    if oldfd == newfd {
-        return newfd;
-    }
-
-    let dup2_result = FN_DUP2(oldfd, newfd);
-
-    if dup2_result == -1 {
-        dup2_result
-    } else {
-        let (Ok(result) | Err(result)) =
-            dup(oldfd, dup2_result)
-                .map(|()| dup2_result)
-                .map_err(|fail| match fail {
-                    LayerError::LocalFDNotFound(_) => dup2_result,
-                    _ => fail.into(),
-                });
-
-        trace!("dup2_detour -> result {:#?}", result);
-        result
-    }
-}
-
-#[cfg(target_os = "linux")]
-#[hook_guard_fn]
-pub(super) unsafe extern "C" fn dup3_detour(oldfd: c_int, newfd: c_int, flags: c_int) -> c_int {
-    trace!(
-        "dup3_detour -> oldfd {:#?} | newfd {:#?} | flags {:#?}",
-        oldfd,
-        newfd,
-        flags
-    );
-
-    let dup3_result = FN_DUP3(oldfd, newfd, flags);
-
-    if dup3_result == -1 {
-        dup3_result
-    } else {
-        let (Ok(result) | Err(result)) =
-            dup(oldfd, dup3_result)
-                .map(|()| dup3_result)
-                .map_err(|fail| match fail {
-                    LayerError::LocalFDNotFound(_) => dup3_result,
-                    _ => fail.into(),
-                });
-
-        trace!("dup3_detour -> result {:#?}", result);
-        result
-    }
-}
 /// Turns the raw pointer parameters into Rust types and calls `ops::getaddrinfo`.
 ///
 /// # Warning:
@@ -438,9 +339,6 @@ pub(crate) unsafe fn enable_socket_hooks(interceptor: &mut Interceptor, enabled_
         FN_CONNECT
     );
 
-    let _ = replace!(interceptor, "fcntl", fcntl_detour, FnFcntl, FN_FCNTL);
-    let _ = replace!(interceptor, "dup", dup_detour, FnDup, FN_DUP);
-    let _ = replace!(interceptor, "dup2", dup2_detour, FnDup2, FN_DUP2);
 
     let _ = replace!(
         interceptor,
@@ -474,9 +372,7 @@ pub(crate) unsafe fn enable_socket_hooks(interceptor: &mut Interceptor, enabled_
             accept4_detour,
             FnAccept4,
             FN_ACCEPT4
-        );
-
-        let _ = replace!(interceptor, "dup3", dup3_detour, FnDup3, FN_DUP3);
+        );        
     }
 
     let _ = replace!(interceptor, "accept", accept_detour, FnAccept, FN_ACCEPT);
