+++ mirrord-layer/src/socket/ops.rs
@@ -306,25 +306,6 @@ pub(super) fn accept(
     Ok(new_fd)
 }
 
-pub(super) fn fcntl(orig_fd: c_int, cmd: c_int, fcntl_fd: i32) -> Result<(), LayerError> {
-    match cmd {
-        libc::F_DUPFD | libc::F_DUPFD_CLOEXEC => dup(orig_fd, fcntl_fd),
-        _ => Ok(()),
-    }
-}
-
-pub(super) fn dup(fd: c_int, dup_fd: i32) -> Result<(), LayerError> {
-    let dup_socket = SOCKETS
-        .lock()?
-        .get(&fd)
-        .ok_or(LayerError::LocalFDNotFound(fd))?
-        .clone();
-
-    SOCKETS.lock()?.insert(dup_fd as RawFd, dup_socket);
-
-    Ok(())
-}
-
 /// Retrieves the result of calling `getaddrinfo` from a remote host (resolves remote DNS),
 /// converting the result into a `Box` allocated raw pointer of `libc::addrinfo` (which is basically
 /// a linked list of such type).
